Allow overflow to fix issue #40
Allow overflow to fix issue #40 

Honestly I'm doing this pull request straight from github, without testing it.  
I have more experience on backend stuff and python, not much in the frontend or typescript. I would have to learn and figure out how to test it. And  I plan to do, WHEN I have more time, but meanwhile this guesswork change looks like the correct thing to do based in the workaround posted. 
If someone can also review or test it I apreciate.!